refactoring a bit menu composer in order to define a menu composer type and use config file instead hard coded entry
