+++ src/Http/Composers/Sidebar.php
@@ -28,7 +28,7 @@ use Seat\Web\Exceptions\PackageMenuBuilderException;
  * Class User
  * @package Seat\Web\Http\Composers
  */
-class Sidebar
+class Sidebar extends AbstractMenu
 {
 
     /**
@@ -37,6 +37,18 @@ class Sidebar
     public function __construct()
     {
         //
+    }
+
+    /**
+     * Return required keys in menu structure
+     *
+     * @return array
+     */
+    public function getRequiredKeys() : array
+    {
+        return [
+            'name', 'route_segment'
+        ];
     }
 
     /**
@@ -51,117 +63,7 @@ class Sidebar
 
         // Home. This menu item declares the menu and
         // sets it as an array of arrays.
-        $menu = [
-            [
-                'name'          => trans('web::seat.home'),
-                'icon'          => 'fa-home',
-                'route_segment' => 'home',
-                'route'         => 'home'
-            ]
-        ];
-
-        // Key Management
-        array_push($menu, [
-
-            'name'          => trans('web::seat.api_key_management'),
-            'icon'          => 'fa-key',
-            'route_segment' => 'api-key',
-            'entries'       => [
-
-                [   // Add Api Key
-                    'name'  => trans('web::seat.add_api_key'),
-                    'icon'  => 'fa-plus',
-                    'route' => 'api.key'
-                ],
-                [
-                    'name'  => trans('web::seat.list_keys'),
-                    'icon'  => 'fa-list',
-                    'route' => 'api.key.list'
-                ],
-                [
-                    'name'  => trans('web::seat.people_groups'),
-                    'icon'  => 'fa-group',
-                    'route' => 'people.list'
-                ]
-            ]
-        ]);
-
-        // Corporation
-        array_push($menu, [
-            'name'          => trans_choice('web::seat.corporation', 2),
-            'icon'          => 'fa-building',
-            'route_segment' => 'corporation',
-            'entries'       => [
-                [
-                    'name'  => trans('web::seat.all_corp'),
-                    'icon'  => 'fa-group',
-                    'route' => 'corporation.list'
-                ]
-            ]
-        ]);
-
-        // Character
-        array_push($menu, [
-            'name'          => trans_choice('web::seat.character', 2),
-            'icon'          => 'fa-user',
-            'route_segment' => 'character',
-            'entries'       => [
-                [
-                    'name'  => trans('web::seat.all_char'),
-                    'icon'  => 'fa-group',
-                    'route' => 'character.list'
-                ],
-                [
-                    'name'  => trans('web::seat.mail_timeline'),
-                    'icon'  => 'fa-envelope',
-                    'route' => 'character.view.mail.timeline'
-                ]
-            ]
-        ]);
-
-        // Configuration
-        if (auth()->user()->hasSuperuser()) {
-
-            array_push($menu, [
-                'name'          => trans('web::seat.configuration'),
-                'icon'          => 'fa-cogs',
-                'route_segment' => 'configuration',
-                'entries'       => [
-
-                    [   // Access
-                        'name'  => trans('web::seat.access'),
-                        'icon'  => 'fa-shield',
-                        'route' => 'configuration.access.roles'
-                    ],
-                    [   // Import
-                        'name'  => trans('web::seat.import'),
-                        'icon'  => 'fa-upload',
-                        'route' => 'configuration.import.list'
-                    ],
-                    [   // Users
-                        'name'  => trans_choice('web::seat.user', 2),
-                        'icon'  => 'fa-user',
-                        'route' => 'configuration.users'
-                    ],
-                    [   // Schedule
-                        'name'  => trans('web::seat.schedule'),
-                        'icon'  => 'fa-clock-o',
-                        'route' => 'configuration.schedule'
-                    ],
-                    [   // SeAT Setting
-                        'name'  => trans('web::seat.settings'),
-                        'icon'  => 'fa-cog',
-                        'route' => 'seat.settings.view'
-                    ],
-                    [   // Security
-                        'name'  => trans('web::seat.security_logs'),
-                        'icon'  => 'fa-list',
-                        'route' => 'configuration.security.logs'
-                    ],
-
-                ]
-            ]);
-        }
+        $menu = [];
 
         // Load any menus from any registered packages
         $package_menus = config('package.sidebar');
@@ -173,106 +75,6 @@ class Sidebar
                 array_push($menu, $prepared_menu);
         }
 
-        array_push($menu, [
-            'name'          => trans('web::seat.other'),
-            'icon'          => 'fa-circle',
-            'route_segment' => 'other',
-        ]);
-
         $view->with('menu', $menu);
-    }
-
-    /**
-     * Load menus from any registered plugins.
-     *
-     * This may end up being quite a complex method, as we need
-     * to validate a lot of the menu structure that is set
-     * out.
-     *
-     * Packages should register menu items in a config file,
-     * loaded in a ServiceProvider's register() method in the
-     * 'package.sidebar' namespace. The structure of these
-     * menus can be seen in the SeAT Wiki.
-     *
-     * @param $package_name
-     * @param $menu_data
-     *
-     * @return array
-     * @throws \Seat\Web\Exceptions\PackageMenuBuilderException
-     */
-    public function load_plugin_menu($package_name, $menu_data)
-    {
-
-        // Validate the package menu
-        $this->validate_menu($package_name, $menu_data);
-
-        // Check if the current user has the permission
-        // required to see the menu
-        if (isset($menu_data['permission']))
-            if (!auth()->user()->has($menu_data['permission']))
-                return;
-
-        return $menu_data;
-    }
-
-    /**
-     * The actual menu valiation logic.
-     *
-     * @param $package_name
-     * @param $menu_data
-     *
-     * @throws \Seat\Web\Exceptions\PackageMenuBuilderException
-     */
-    public function validate_menu($package_name, $menu_data)
-    {
-
-        if (!is_string($package_name))
-            throw new PackageMenuBuilderException(
-                'Package root menu items should be named by string type');
-
-        if (!is_array($menu_data))
-            throw new PackageMenuBuilderException(
-                'Package menu data should be defined in an array');
-
-        if (!array_key_exists('name', $menu_data))
-            throw new PackageMenuBuilderException(
-                'Root menu must define a name');
-
-        if (!array_key_exists('route_segment', $menu_data))
-            throw new PackageMenuBuilderException(
-                'Root menu must define a route segement');
-
-        // Check if we have sub entries. If not, we have to
-        // check if we have a route defined in the parent
-        // menu.
-        if (!array_key_exists('entries', $menu_data))
-            if (!array_key_exists('route', $menu_data))
-                throw new PackageMenuBuilderException(
-                    'A parent or sub-menu route is required.');
-
-        // Ensure that the entries is actually an array
-        if (array_key_exists('entries', $menu_data) && !is_array($menu_data['entries']))
-            throw new PackageMenuBuilderException(
-                'Root menu must define entries');
-
-        // Loop over the sub menu entries, validating the
-        // required fields
-        if (isset($menu_data['entries'])) {
-
-            foreach ($menu_data['entries'] as $entry) {
-
-                if (!array_key_exists('name', $entry))
-                    throw new PackageMenuBuilderException(
-                        'A sub menu entry failed to define a name');
-
-                if (!array_key_exists('icon', $entry))
-                    throw new PackageMenuBuilderException(
-                        'A sub menu entry failed to define an icon');
-
-                if (!array_key_exists('route', $entry))
-                    throw new PackageMenuBuilderException(
-                        'A sub menu entry failed to define a route');
-            }
-        }
     }
 }
