+++ src/resources/views/includes/sidebar.blade.php
@@ -46,7 +46,11 @@
           <li class="treeview {{ Request::segment(1) === $entry['route_segment'] ? 'active' : null }}">
             <a href="#">
               <i class="fa {{ $entry['icon'] }}"></i>
+              @if (array_key_exists('label', $entry))
+              <span>{{ trans($entry['label']) }}</span> <i class="fa fa-angle-left pull-right"></i>
+              @else
               <span>{{ $entry['name'] }}</span> <i class="fa fa-angle-left pull-right"></i>
+              @endif
             </a>
             <ul class="treeview-menu">
 
@@ -54,7 +58,12 @@
 
                 <li class="{{ isset($item['route']) ? (Request::url() === route($item['route']) ? 'active' : null) : null }}">
                   <a href="{{ isset($item['route']) ? route($item['route']) : '#' }}">
-                    <i class="fa {{ $item['icon'] or 'fa-circle-o' }}"></i> {{ $item['name'] }}</a>
+                    @if (array_key_exists('label', $item))
+                    <i class="fa {{ $item['icon'] or 'fa-circle-o' }}"></i> {{ trans($item['label']) }}
+                    @else
+                    <i class="fa {{ $item['icon'] or 'fa-circle-o' }}"></i> {{ $item['name'] }}
+                    @endif
+                  </a>
                 </li>
 
               @endforeach
@@ -66,7 +75,11 @@
 
           <li class="{{ Request::segment(1) === $entry['route_segment'] ? 'active' : null }}">
             <a href="{{ isset($entry['route']) ? route($entry['route']) : '#' }}">
+              @if (array_key_exists('label', $entry))
+              <i class="fa {{ $entry['icon'] }}"></i> <span>{{ trans($entry['label']) }}</span>
+              @else
               <i class="fa {{ $entry['icon'] }}"></i> <span>{{ $entry['name'] }}</span>
+              @endif
             </a>
           </li>
